+++ lua/obsidian/types.lua
@@ -52,7 +52,6 @@
 ---@field unique_note? obsidian.config.UniqueNoteOpts
 ---@field sync? obsidian.config.SyncOpts
 ---@field slides? obsidian.config.SlidesOpts
----@field bookmarks? obsidian.config.BookmarksOpts
 
 ---@class obsidian.config.Internal
 ---@field workspaces obsidian.workspace.WorkspaceSpec[]
@@ -83,6 +82,5 @@
 ---@field unique_note obsidian.config.UniqueNoteOpts
 ---@field sync obsidian.config.SyncOpts
 ---@field slides obsidian.config.SlidesOpts
----@field bookmarks obsidian.config.BookmarksOpts
 
 ---@alias obsidian.config.NewNotesLocation "current_dir" | "notes_subdir"